+++ src/helpers/swapTransactions.ts
@@ -1,11 +1,8 @@
-import { incrementErrorCount, performRpcQuery, selectNodeProviders } from './queryNodes';
 import { SwapObject, TransactionResult, RPCResponse, Asset } from '@/types';
-import { CHAIN_ENDPOINTS, DELAY_BETWEEN_NODE_ATTEMPTS, MAX_NODES_PER_QUERY } from '@/constants';
-import { createOfflineSignerFromMnemonic } from './dataHelpers/wallet';
-import { delay } from './timer';
+import { CHAIN_ENDPOINTS } from '@/constants';
 import { getValidFeeDenom } from './feeDenom';
-import { getSessionToken } from './dataHelpers';
-import { getSigningSymphonyClient, symphony } from '@orchestra-labs/symphonyjs';
+import { symphony } from '@orchestra-labs/symphonyjs';
+import { querySymphony } from './querySymphonyNodes';
 
 const { swapSend } = symphony.market.v1beta1.MessageComposer.withTypeUrl;
 
@@ -19,68 +16,6 @@ export const isValidSwap = ({
   const result = !sendAsset.isIbc && !receiveAsset.isIbc && sendAsset.denom !== receiveAsset.denom;
 
   return result;
-};
-
-// TODO: merge in with queryNodes.  do not need separate signer if endpoint and message object are used
-const queryWithRetry = async ({
-  endpoint,
-  walletAddress,
-  messages = [],
-  feeDenom,
-  simulateOnly = false,
-}: {
-  endpoint: string;
-  walletAddress: string;
-  messages?: any[];
-  feeDenom: string;
-  simulateOnly?: boolean;
-}): Promise<any> => {
-  const providers = selectNodeProviders();
-  console.log('Selected node providers:', providers);
-
-  let numberAttempts = 0;
-
-  while (numberAttempts < MAX_NODES_PER_QUERY) {
-    for (const provider of providers) {
-      try {
-        const queryMethod = provider.rpc;
-        console.log(`Querying node ${queryMethod} with endpoint: ${endpoint}`);
-
-        const sessionToken = getSessionToken();
-        if (!sessionToken) {
-          console.error('Error- getSessionTokenFailed');
-          return;
-        }
-        const offlineSigner = await createOfflineSignerFromMnemonic(sessionToken.mnemonic || '');
-
-        const client = await getSigningSymphonyClient({
-          rpcEndpoint: queryMethod,
-          signer: offlineSigner,
-        });
-
-        const result = await performRpcQuery(
-          client,
-          walletAddress,
-          messages,
-          feeDenom,
-          simulateOnly,
-        );
-        return result;
-      } catch (error) {
-        incrementErrorCount(provider.rpc);
-        console.error('Error querying node:', error);
-      }
-      numberAttempts++;
-
-      if (numberAttempts >= MAX_NODES_PER_QUERY) {
-        break;
-      }
-
-      await delay(DELAY_BETWEEN_NODE_ATTEMPTS);
-    }
-  }
-
-  throw new Error(`All node query attempts failed after ${MAX_NODES_PER_QUERY} attempts.`);
 };
 
 export const swapTransaction = async (
@@ -109,7 +44,7 @@ export const swapTransaction = async (
       swapObject.sendObject.symphonyAssets,
     );
     console.log('Swap fee denom:', feeDenom);
-    const response = await queryWithRetry({
+    const response = await querySymphony({
       endpoint,
       walletAddress: fromAddress,
       messages,
@@ -173,7 +108,7 @@ export const multiSwapTransaction = async (
       swapObjects[0].sendObject.denom,
       swapObjects[0].sendObject.symphonyAssets,
     );
-    const response = await queryWithRetry({
+    const response = await querySymphony({
       endpoint,
       walletAddress: fromAddress,
       messages,
